+++ changelog.md
@@ -4,6 +4,7 @@
 * Including a "details" message in `DuplicateCloudEventException` that adds more details on why this happens (which index etc). This is especially useful
   if you're creating custom, unique, indexes over the events and a write fail due to a duplicate cloud event.
 * Upgraded to Kotlin 1.3.40
+* Upgraded project-reactor to 3.4.2 (previously 3.4.0 was used)
 
 ## Changelog 0.7.0 (2021-01-31)